+++ crates/zng-view/src/lib.rs
@@ -1630,6 +1630,10 @@ impl Api for App {
         self.with_window(id, |w| w.drag_move(), || ())
     }
 
+    fn drag_resize(&mut self, id: WindowId, direction: zng_view_api::window::ResizeDirection) {
+        self.with_window(id, |w| w.drag_resize(direction), || ())
+    }
+
     fn open_title_bar_context_menu(&mut self, id: WindowId, position: DipPoint) {
         self.with_window(id, |w| w.open_title_bar_context_menu(position), || ())
     }